// templates/typescript/samples/sample-assistant/src/models/stateProperties.ts
/**
 * Copyright(c) Microsoft Corporation.All rights reserved.
 * Licensed under the MIT License.
 */

export enum StateProperties {
    DispatchResult = 'dispatchResult',
    GeneralResult = 'generalResult',
    PreviousBotResponse = 'previousBotResponse',
    Location = 'location',
    TimeZone = 'timeZone'
}